requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.22
pydantic>=2.0
requests>=2.28
python-dotenv>=1.0
